requirements.txt
streamlit
pandas
pytest
flake8
Pillow
